move iteration code to separate file
